refactor(tests): move gemini token usage tests to test_vertex_and_google_ai_studio_gemini.py
